backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.6.0
python-multipart>=0.0.6
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
slowapi
pydantic>=2.0
pydantic-settings>=2.0
pydicom>=2.4.0
Pillow>=10.0.0
aiofiles>=23.0.0